comment component refactoring
