+++ src/components/posts/PostsPage.js
@@ -10,7 +10,7 @@ const PostsPage = ({ role }) => {
   const [post, setPost] = useState({ content: '' })
   const [postResponse, setPostResponse] = useState('')
   const [posts, setPosts] = useState([])
-  console.log(posts)
+
   useEffect(() => {
     client.get('/posts').then((res) => {
       setPosts(res.data.data.posts)
@@ -25,8 +25,8 @@ const PostsPage = ({ role }) => {
         setPostResponse(res.data)
         setPosts((posts) => [res.data.data.post, ...posts])
       })
-      .catch((data) => {
-        console.log(data)
+      .catch((err) => {
+        console.log(err.message, 'Invalid post')
       })
     setPost(() => ({ content: '' }))
   }
