+++ src/components/posts/style.css
@@ -112,7 +112,7 @@ width:100%
   justify-content: flex-end;
 }
 
-.single-comment {
+.first-comment {
   font-size: 10pt;
   margin-top:1em;
 }
